requirements.txt
amqp==5.2.0
asgiref==3.7.2
billiard==4.2.0
celery==5.3.5
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==41.0.5
Deprecated==1.2.14
Django==4.2.7
django-cors-headers==4.3.1
django-oauth-toolkit==2.3.0
djangorestframework==3.14.0
dnspython==2.4.2
eventlet==0.33.3
ez-setup==0.9
greenlet==3.0.1
gunicorn==21.2.0
idna==3.4
jwcrypto==1.5.0
kombu==5.3.3
numpy==1.26.1
oauthlib==3.2.2
opencv-python-headless==4.8.1.78
packaging==23.2
prompt-toolkit==3.0.40
psycopg2-binary==2.9.9
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
redis==5.0.1
requests==2.31.0
setuptools==68.2.2
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.0.7
vine==5.1.0
wcwidth==0.2.9
wheel==0.37.1
whitenoise==6.6.0
wrapt==1.16.0
